Do a pass over schemas, change some things up.
So first off, Plotinus is going to be disabled by default for release 12. The reasoning behind this is it has the tendency to break some apps, like Firefox. For release 13, I'll have a load of applications that'll be explicitly enabled, but I'll be moving to a whitelist system instead.

I updated the wallpaper to the new default. For Nautilus, I set recursive-search to local-only so it's not trying to search across every possible device you have and causing it to choke on lower-end systems. I enabled Nautilus show-image-thumbnails for all devices, so that will include network shares ala Samba.